requirements.txt
cloudshell-cp-azure>=1.6.0,<1.7.0
